[useMenu] refactors menuItem type interfaces to new file to resolve circular dependency problem
