+++ packages/mui-base/src/useMenu/menuItem.types.ts
@@ -1,0 +1,11 @@
+export interface MenuItemMetadata {
+    id: string;
+    disabled: boolean;
+    label?: string;
+    ref: React.RefObject<HTMLElement>;
+}
+
+export interface MenuItemState {
+    disabled: boolean;
+    highlighted: boolean;
+  }
